nix_bindings_expr_sys

Type Alias PrimOpFun

pub type PrimOpFun = Option<unsafe extern "C" fn(user_data: *mut c_void, context: *mut c_context, state: *mut EvalState, args: *mut *mut value, ret: *mut value)>;

@brief Function pointer for primops @ingroup primops

When you want to return an error, call nix_set_err_msg(context, NIX_ERR_UNKNOWN, “your error message here”).

@param[in] user_data Arbitrary data that was initially supplied to nix_alloc_primop @param[out] context Stores error information. @param[in] state Evaluator state @param[in] args list of arguments. Note that these can be thunks and should be forced using nix_value_force before use. @param[out] ret return value @see nix_alloc_primop, nix_init_primop

Aliased Type

pub enum PrimOpFun {
    None,
    Some(unsafe extern "C" fn(*mut c_void, *mut c_context, *mut EvalState, *mut *mut value, *mut value)),
}

Variants

None

No value.

Some(unsafe extern "C" fn(*mut c_void, *mut c_context, *mut EvalState, *mut *mut value, *mut value))

Some value of type T.